+++ es-server/src/main/java/app/repositories/TripsRepositoryJpa.java
@@ -12,32 +12,44 @@ import javax.persistence.TypedQuery;
 import javax.transaction.Transactional;
 import java.util.List;
 
-@Repository
+@Repository("TRIPS.JPA")
 @Transactional
 @Primary
-public class TripsRepositoryJpa {
+public class TripsRepositoryJpa
+  extends AbstractEntityRepositoryJpa<Trip> {
 
   @Autowired
   @PersistenceContext
   EntityManager entityManager;
 
+  public TripsRepositoryJpa() {
+    super(Trip.class);
+  }
+
+  @Override
   public List<Trip> findAll() {
     TypedQuery<Trip> query = this.entityManager.createQuery("select t from Trip t", Trip.class);
     return query.getResultList();
   }
 
+  @Override
   public Trip findById(long id) {
     return entityManager.find(Trip.class, id);
   }
 
+  @Override
   //Saves new trip or updates trip
   public Trip save(Trip trip) {
     return entityManager.merge(trip);
   }
 
-  public Trip deleteById(long id) {
+  @Override
+  public boolean deleteById(long id) {
     Trip trip = findById(id);
-    entityManager.remove(trip);
-    return null;
+    if(trip != null){
+      entityManager.remove(trip);
+      return true;
+    }
+    return false;
   }
 }
